compute/requirements.txt
mock==4.0.3
pika==1.3.0
pytest==7.1.2
Pillow==9.2.0
numpy==1.23.1
